LayoutTests/platform/gtk/media/controls-drag-timebar-expected.txt

Test that drag the timebar thumb causes seeks.

RUN(video.autoplay = true)
RUN(video.src = 'content/test.mp4')
EVENT(playing)
EVENT(seeked)
Time: 2.4
EVENT(seeked)
Time: 3.0
EVENT(seeked)
Time: 3.6
END OF TEST
